UI Updates
Added build menu with basic open/close functionality, cost/effect descriptions, and a buy button that is currently not connected to anything.
Changed background color to be green.
Added a next day button that is not connected to anything.
Added a quit game button so I don't have to Alt-F4 out of the game all the time.
Primary game space is now complete.
All sprint tasks are now complete.