Resource and Building Updates
There are now multiple types of buildings and resources.
Resources are automatically generated upon starting the game.
Buildings and resources can now be destroyed.
Buildings can be purchased and placed from the build menu.

Problems:
Resources are not supposed to generate on top of each other or on top of buildings, but this is happening anyways.
The attack math has been started, but will require some changes to fully get working. Game manager should have a list of all current building objects with actual object references and should have a way to determine exactly what type of building each one is.